pre/f1/pyproject.toml
[project]
name = "f1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "pytest-playwright>=0.5.2",
    "requests>=2.32.3",
    "rich>=13.9.3",
    "typer>=0.12.5",
]
